% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmeteo.R
\docType{data}
\name{dmeteo}
\alias{dmeteo}
\title{1926 days of weather station observations.}
\format{
A data frame with 46220 rows and 7 variables:
\describe{
  \item{fecha}{Date}
  \item{hora}{Character}
  \item{tmed}{Temperature in Celsius scale.}
  \item{hrmed}{Relative humidity (\%).}
  \item{vvmed}{Wind speed in m/s.}
  \item{radmed}{Numeric vector of solar irradiation in W/m2.}
  \item{FechaHora}{POSIXct}
}
}
\usage{
dmeteo
}
\description{
1926 days of weather station observations.
}
\examples{
dmeteo
}
\keyword{datasets}
