% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0.R
\docType{package}
\name{ET.PenmanMonteith}
\alias{ET.PenmanMonteith}
\title{ET.PenmanMonteith: A package for calculate daily Penman-Monteith Evapotranspiration from hourly data.}
\description{
This package is a shell for the function \code{\link[Evapotranspiration]{ET.PenmanMonteith}} from \code{Evapotranspiration} package written by \href{https://cran.r-project.org/web/packages/Evapotranspiration/index.html}{Danlu Guo}.
}
