% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/et0.R
\name{et0}
\alias{et0}
\title{Get daily Penman-Monteith Evapotranspiration from hourly data}
\usage{
et0(dates, temp, hr, uz, rs, lat, elev, crop = "short", msg = "no")
}
\arguments{
\item{dates}{POSIXct dates vector.}

\item{temp}{numeric vector of temperature in Celsius scale.}

\item{hr}{numeric vector of relative humidity (\%).}

\item{uz}{numeric vector of wind speed in m/s.}

\item{rs}{numeric vector of solar irradiation in W/m2}

\item{lat}{latitude.}

\item{elev}{numeric elevation in mts.}

\item{crop}{character "short" (default) of "tall". See \code{\link[Evapotranspiration]{ET.PenmanMonteith}}.}

\item{msg}{character "yes" or "no".  See param "message" in \code{\link[Evapotranspiration]{ET.PenmanMonteith}}.}
}
\value{
dataframe with columns \code{date} and \code{et0} containing the daily Penman-Monteith Reference Evapotranspiration
}
\description{
Get daily Penman-Monteith Evapotranspiration from hourly data
}
\examples{
data(meteo)
ET.PenmanMonteith::et0(
  dates = meteo$dates,
  temp  = meteo$temp,
  hr    = meteo$higro,
  uz    = meteo$anemo,
  rs    = meteo$piro,
  lat   = 38.04371,
  elev  = 313) -> et0.df
head(et0.df)
}
