% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo.R
\docType{data}
\name{meteo}
\alias{meteo}
\title{125 days of weather station observations.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{dates}{Hourly POSIXct.}
  \item{temp}{Temperature in Celsius scale.}
  \item{piro}{Numeric vector of solar irradiation in W/m2.}
  \item{higro}{Relative humidity (\%).}
  \item{pluv}{Precipitation in mm.}
  \item{anemo}{Wind speed in m/s.}
  \item{baro}{Pressure in mb.}
}
}
\usage{
meteo
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\examples{
meteo
}
\keyword{datasets}
